+++ packages/antwerp-ui/react-components/src/lib/atoms/input/text-field/TextField.stories.jsx
@@ -24,6 +24,7 @@ export default {
     required: false,
     charCounter: false,
     charCountText: '%count% / %max%',
+    maxLengthBlocksInput: false,
     addOn: { type: 'icon', content: 'phone', placement: 'left' },
     id: 'textfield-story'
   },
@@ -126,6 +127,14 @@ export default {
         type: { summary: 'integer' }
       },
       description: 'The `maxLength` attribute sets a character limitation for the value of the `TextField`.'
+    },
+    maxLengthBlocksInput: {
+      control: { type: 'boolean' },
+      table: {
+        type: { sunnary: 'boolean' }
+      },
+      description:
+        "When the `maxLength` prop is set, the `maxLengthBlocksInput` prop blocks the user from typing more characters than the `maxLength` prop's value."
     },
     charCountText: {
       control: { type: 'text' },